feat(@core/classes): re-structure classes
