+++ src/app/core/classes/ItalianAppetizer.ts
@@ -1,0 +1,8 @@
+import { Appetizer } from '../interfaces/appetizer';
+
+/**
+ * Class representing a italian appetizer.
+ */
+export class ItalianAppetizer implements Appetizer {
+  name = 'italianAppetizer';
+}
