+++ src/app/core/classes/ItalianDessert.ts
@@ -1,0 +1,8 @@
+import { Dessert } from '../interfaces/dessert';
+
+/**
+ * Class representing a italian dessert.
+ */
+export class ItalianDessert implements Dessert {
+  name = 'Dessert';
+}
