+++ src/app/core/classes/ItalianMainCourse.ts
@@ -1,0 +1,8 @@
+import { MainCourse } from '../interfaces/mainCourse';
+
+/**
+ * Class representing a italian mainCourse.
+ */
+export class ItalianMainCourse implements MainCourse {
+  name = 'italianMainCourse';
+}
