+++ src/app/core/classes/MexicanAppetizer.ts
@@ -1,0 +1,8 @@
+import { Appetizer } from '../interfaces/appetizer';
+
+/**
+ * Class representing a mexican appetizer.
+ */
+export class MexicanAppetizer implements Appetizer {
+  name = 'mexicanAppetizer';
+}
