+++ src/app/core/classes/MexicanDessert.ts
@@ -1,0 +1,8 @@
+import { Dessert } from '../interfaces/dessert';
+
+/**
+ * Class representing a mexican dessert.
+ */
+export class MexicanDessert implements Dessert {
+  name = 'mexicanDessert';
+}
