+++ src/app/core/classes/MexicanMainCourse.ts
@@ -1,0 +1,8 @@
+import { MainCourse } from '../interfaces/mainCourse';
+
+/**
+ * Class representing a mexican main course.
+ */
+export class MexicanMainCourse implements MainCourse {
+  name = 'mexicanMainCourse';
+}
