+++ src/app/core/classes/mexican-restaurant-factory.ts
@@ -1,14 +1,14 @@
-import { Appetizer, Dessert, MainCourse } from '../interfaces/dishes';
+import { Appetizer } from '../interfaces/appetizer';
+import { Dessert } from '../interfaces/dessert';
+import { MainCourse } from '../interfaces/mainCourse';
 import { Restaurants } from '../interfaces/restaurant';
-import {
-  MexicanDessert,
-  MexicanMainCourse,
-  mexicanAppetizer,
-} from './mexicanRestaurant';
+import { MexicanAppetizer } from './MexicanAppetizer';
+import { MexicanDessert } from './MexicanDessert';
+import { MexicanMainCourse } from './MexicanMainCourse';
 
 export class MexicanRestaurantFactory implements Restaurants {
   createAppetizer(): Appetizer {
-    return mexicanAppetizer;
+    return MexicanAppetizer;
   }
   createDessert(): Dessert {
     return MexicanMainCourse;
