+++ src/app/core/interfaces/appetizer.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface representing an appetizer dish.
+ */
+
+export interface Appetizer {
+  name: string;
+}
