+++ src/app/core/interfaces/dessert.ts
@@ -1,0 +1,6 @@
+/**
+ * Interface representing a dessert dish.
+ */
+export interface Dessert {
+  name: string;
+}
