--- src/app/core/interfaces/dishes.ts
@@ -1,21 +1,0 @@
-/**
- * Interface representing an appetizer dish.
- */
-
-export interface Appetizer {
-  name: string;
-}
-
-/**
- * Interface representing a mainCourse dish.
- */
-export interface MainCourse {
-  name: string;
-}
-
-/**
- * Interface representing a dessert dish.
- */
-export interface Dessert {
-  name: string;
-}
