+++ src/app/core/interfaces/mainCourse.ts
@@ -1,0 +1,6 @@
+/**
+ * Interface representing a mainCourse dish.
+ */
+export interface MainCourse {
+  name: string;
+}
